[FE] styles : API 영역 및 Type 처리 @shared 디렉토리로 이동
