+++ frontend/src/service/@shared/hooks/queries/profile/useGetMyReviewForms.ts
@@ -1,10 +1,10 @@
 import { useQuery, UseQueryOptions } from 'react-query';
 
 import { ErrorResponse } from 'service/@shared/types/index';
-import { GetMyReviewFormsResponse } from 'service/community/types';
+import { GetMyReviewFormsResponse } from 'service/@shared/types/profile';
 
+import myReviewAPI from 'service/@shared/api/profile';
 import { QUERY_KEY } from 'service/@shared/constants';
-import myReviewAPI from 'service/community/api';
 
 function useGetMyReviewForms(
   queryOptions?: UseQueryOptions<GetMyReviewFormsResponse, ErrorResponse>,
