+++ frontend/src/service/@shared/hooks/queries/profile/useGetMyReviews.ts
@@ -1,10 +1,10 @@
 import { useQuery, UseQueryOptions } from 'react-query';
 
 import { ErrorResponse } from 'service/@shared/types/index';
-import { GetMyReviewsResponse } from 'service/community/types';
+import { GetMyReviewsResponse } from 'service/@shared/types/profile';
 
+import myReviewAPI from 'service/@shared/api/profile';
 import { QUERY_KEY } from 'service/@shared/constants';
-import myReviewAPI from 'service/community/api';
 
 function useGetMyReviews(queryOptions?: UseQueryOptions<GetMyReviewsResponse, ErrorResponse>) {
   return useQuery<GetMyReviewsResponse, ErrorResponse>(
