+++ frontend/src/service/@shared/hooks/queries/review/useDeleteReview.ts
@@ -1,14 +1,14 @@
 import { useMutation, useQueryClient } from 'react-query';
 
-import { UseCustomMutationOptions } from 'service/review/types';
+import { UseCustomMutationOptions } from 'service/@shared/types/review';
 
+import reviewAPI from 'service/@shared/api/review';
 import { QUERY_KEY } from 'service/@shared/constants';
-import reviewAPI from 'service/review/api';
 
-function useDeleteReviewForm(mutationOptions?: UseCustomMutationOptions<null>) {
+function useDeleteReview(mutationOptions?: UseCustomMutationOptions<null>) {
   const queryClient = useQueryClient();
 
-  return useMutation(reviewAPI.deleteReviewForm, {
+  return useMutation(reviewAPI.deleteReview, {
     onSuccess: () => {
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW]);
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW_FORM]);
@@ -17,4 +17,4 @@ function useDeleteReviewForm(mutationOptions?: UseCustomMutationOptions<null>) {
   });
 }
 
-export default useDeleteReviewForm;
+export default useDeleteReview;
