+++ frontend/src/service/@shared/hooks/queries/review/useDeleteReviewForm.ts
@@ -1,14 +1,14 @@
 import { useMutation, useQueryClient } from 'react-query';
 
-import { UseCustomMutationOptions } from 'service/review/types';
+import { UseCustomMutationOptions } from 'service/@shared/types/review';
 
+import reviewAPI from 'service/@shared/api/review';
 import { QUERY_KEY } from 'service/@shared/constants';
-import reviewAPI from 'service/review/api';
 
-function useDeleteReview(mutationOptions?: UseCustomMutationOptions<null>) {
+function useDeleteReviewForm(mutationOptions?: UseCustomMutationOptions<null>) {
   const queryClient = useQueryClient();
 
-  return useMutation(reviewAPI.deleteReview, {
+  return useMutation(reviewAPI.deleteReviewForm, {
     onSuccess: () => {
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW]);
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW_FORM]);
@@ -17,4 +17,4 @@ function useDeleteReview(mutationOptions?: UseCustomMutationOptions<null>) {
   });
 }
 
-export default useDeleteReview;
+export default useDeleteReviewForm;
