+++ frontend/src/service/@shared/hooks/queries/review/useUpdateReviewForm.ts
@@ -1,9 +1,9 @@
 import { useMutation, useQueryClient } from 'react-query';
 
-import { UpdateReviewFormResponse, UseCustomMutationOptions } from 'service/review/types';
+import { UpdateReviewFormResponse, UseCustomMutationOptions } from 'service/@shared/types/review';
 
+import reviewAPI from 'service/@shared/api/review';
 import { QUERY_KEY } from 'service/@shared/constants';
-import reviewAPI from 'service/review/api';
 
 function useUpdateReviewForm(mutationOptions?: UseCustomMutationOptions<UpdateReviewFormResponse>) {
   const queryClient = useQueryClient();
