+++ frontend/src/service/@shared/hooks/queries/user/useDeleteRefreshToken.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from 'react-query';
 
-import { UseCustomMutationOptions } from 'service/review/types';
+import { UseCustomMutationOptions } from 'service/@shared/types/review';
 
 import { userAPI } from 'service/@shared/api';
 import { QUERY_KEY } from 'service/@shared/constants';
